// src/enum/StockDetailsKeys.enum.ts
export enum StockDetailTypeKey {
  NONE = "NONE",
  OPEN = "OPEN",
  HIGH = "HIGH",
  LOW = "LOW",
  CLOSE = "CLOSE",
  VOLUME = "VOLUME",
}

/**
 * used to map API data keys to normal values
 */
export enum StockDetailTypeValue {
  NONE = "NONE",
  OPEN = "1. open",
  HIGH = "2. high",
  LOW = "3. low",
  CLOSE = "4. close",
  VOLUME = "5. volume",
}