// src/enum/SymbolTypeKeys.enum.ts
export enum SybmolTypeKey {
  NONE = "NONE",
  SYMBOL = "SYMBOL",
  NAME = "NAME",
  TYPE = "TYPE",
  REGION = "REGION",
  MARKET_OPEN = "MARKET_OPEN",
  MARKET_CLOSE = "MARKET_CLOSE",
  TIMEZONE = "TIMEZONE",
  CURRENCY = "CURRENCY",
  MATCH_SCORE = "MATCH_SCORE",
}

export enum SymbolTypeValue {
  NONE = "NONE",
  SYMBOL = "1. symbol",
  NAME = "2. name",
  TYPE = "3. type",
  REGION = "4. region",
  MARKET_OPEN = "5. marketOpen",
  MARKET_CLOSE = "6. marketClose",
  TIMEZONE = "7. timezone",
  CURRENCY = "8. currency",
  MATCH_SCORE = "9. matchScore",
}